Random segfaults should now stop happening! I feel like I should know how to use arrays by now. Oh, that reminds me, TODO: Stop keeping track of that one vector size and just use vector.size(). Also, corners in dungeons built by corruptible_pblind_db are now "corrupted" as they should be. Now I just have to work on the rest of the wall blocks.

+++ src/gui/gui_render.cpp
@@ -6,8 +6,8 @@ void GUI::OnRender()
     canvas.refresh(); 
     TileMatrix tm = canvas.get_matrix();
 
-    for(int i = 0; i < tm.size(); i++) {
-        for(int j = 0; j < tm[i].size(); j++) {
+    for(size_t i = 0; i < tm.size(); i++) {
+        for(size_t j = 0; j < tm[i].size(); j++) {
             drawChr(j, i, tm[i][j].char_count, ascii, screen, tm[i][j].color);
         }
     }
